Interface - v 1.3.0
* Break out Prawn::Icon::Interface into its own file [#27].
* Add simplecov as a development dependency.
* Update rubocop to 0.44.1.
* Bump version to 1.3.0.
