+++ .gitignore
@@ -2,3 +2,4 @@
 Gemfile.lock
 /doc
 .yardoc/
+coverage/
