+++ lib/prawn/icon/version.rb
@@ -8,6 +8,6 @@
 
 module Prawn
   class Icon
-    VERSION = '1.2.0'.freeze
+    VERSION = '1.3.0'.freeze
   end
 end
